requirements.txt
feedparser
beautifulsoup4
requests
openai
pyyaml
